pancheck/client_app/requirements.txt
selenium
webdriver-manager
pandas
openpyxl
customtkinter
cryptography
requests
pillow
